src/backend/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
requests==2.32.3
djangorestframework==3.15.1
drf-yasg==1.21.7
